Move ApplicationPermissions to linera-base.
