+++ linera-core/src/client.rs
@@ -25,7 +25,7 @@ use futures::{
 use linera_base::{
     abi::{Abi, ContractAbi},
     crypto::{CryptoHash, KeyPair, PublicKey},
-    data_types::{Amount, ArithmeticError, BlockHeight, Round, Timestamp},
+    data_types::{Amount, ApplicationPermissions, ArithmeticError, BlockHeight, Round, Timestamp},
     ensure,
     identifiers::{Account, ApplicationId, BytecodeId, ChainId, MessageId, Owner},
     ownership::{ChainOwnership, TimeoutConfig},
@@ -40,9 +40,8 @@ use linera_chain::{
 use linera_execution::{
     committee::{Committee, Epoch, ValidatorName},
     system::{
-        AdminOperation, ApplicationPermissions, OpenChainConfig, Recipient, SystemChannel,
-        SystemOperation, UserData, CREATE_APPLICATION_MESSAGE_INDEX, OPEN_CHAIN_MESSAGE_INDEX,
-        PUBLISH_BYTECODE_MESSAGE_INDEX,
+        AdminOperation, OpenChainConfig, Recipient, SystemChannel, SystemOperation, UserData,
+        CREATE_APPLICATION_MESSAGE_INDEX, OPEN_CHAIN_MESSAGE_INDEX, PUBLISH_BYTECODE_MESSAGE_INDEX,
     },
     Bytecode, ExecutionError, Message, Operation, Query, Response, SystemExecutionError,
     SystemMessage, SystemQuery, SystemResponse, UserApplicationId,
