+++ linera-execution/src/test_utils/system_execution_state.rs
@@ -5,14 +5,14 @@ use crate::{
     applications::ApplicationRegistry,
     committee::{Committee, Epoch},
     execution::UserAction,
-    system::{ApplicationPermissions, SystemChannel},
+    system::SystemChannel,
     ChannelSubscription, ExecutionError, ExecutionRuntimeConfig, ExecutionRuntimeContext,
     ExecutionStateView, OperationContext, ResourceControlPolicy, ResourceController,
     ResourceTracker, TestExecutionRuntimeContext, UserApplicationDescription, UserContractCode,
 };
 use linera_base::{
     crypto::CryptoHash,
-    data_types::{Amount, Timestamp},
+    data_types::{Amount, ApplicationPermissions, Timestamp},
     identifiers::{ApplicationId, ChainDescription, ChainId, Owner},
     ownership::ChainOwnership,
 };
